Abstract all operations into collection methods
