+++ lib/collections/CommentCollectionMethods.js
@@ -1,52 +1,21 @@
 import { check, Maybe } from 'meteor/check'
 
 import comment from '../services/comment'
-import userService from '../services/user'
 import mediaService from '../services/media'
 import { removeReplyByReplyId } from '../services/reply'
 import { getReferenceIdByDocumentId } from './methods/comments'
-import { noOptOptions } from '../noOptOptions'
-
-// TODO: put all these funcs into lib / helper / util libraries that make them re-usable
-const triggerEvent = (name, action, payload) => {
-  const func = Comments.config().onEvent
-
-  if (func) func(name, action, payload)
-}
-
-const verifyContent = content => {
-  check(content, String)
-  return content.trim()
-}
-
-const verifyUserData = (meteorUserId, anonUserData, referenceId) => {
-  check(anonUserData, Match.Maybe(Object))
-  anonUserData = anonUserData || {}
-
-  userService.verifyAnonUserData(anonUserData, referenceId)
-
-  return { anonUserData, userId: (meteorUserId || anonUserData._id) }
-}
+import { noOptOptions } from '../lib/noOptOptions'
+import { triggerEvent } from '../lib/triggerEvent'
+import { verifyUserData, verifyContent } from '../lib/verifyData'
+import createRatingScoreService from '../lib/ratingScore'
 
 /**
- * Add root level methods to the collection.
+ * Add comment methods to the collection.
  *
  * @param {Object} CommentsCollection
  */
 export const addCollectionMethods = (CommentsCollection) => {
-  const getRatingScore = (doc) => {
-    let score = CommentsCollection._calculateAverageRating(doc.starRatings)
-
-    if (score ===  0) {
-      score = doc.likes.length - (doc.dislikes && doc.dislikes.length ? doc.dislikes.length : 0)
-    }
-
-    return score
-  }
-
-  const updateRatingScoreOnDoc = (_id) => CommentsCollection.update({ _id }, {
-    $set: { ratingScore: getRatingScore(CommentsCollection.findOne(_id)) },
-  })
+  const { updateRatingScoreOnDoc } = createRatingScoreService(CommentsCollection)
 
   CommentsCollection.addComment = (referenceId, content, userId, anonData) => {
     const data = verifyUserData(userId, anonData, referenceId)
@@ -209,6 +178,14 @@ export const addCollectionMethods = (CommentsCollection) => {
     _id,
     ...comment.commentStatusSelector(userId)
   })
+
+  CommentsCollection.getCountForReferenceId = (referenceId, userId, anonUserData) => {
+    check(referenceId, String)
+
+    userId = verifyUserData(userId, anonUserData, referenceId).userId
+
+    return CommentsCollection.find(comment.getCommentsSelector(referenceId, userId)).count()
+  }
 
   CommentsCollection.removeReplyByReplyId = (replyId, removeIf) => {
     const doc = CommentsCollection.findOne(
