+++ lib/collections/comments.js
@@ -5,6 +5,7 @@ import timeTickService from '../services/time-tick'
 import userService from '../services/user'
 import comment from '../services/comment'
 import { addCollectionMethods } from './CommentCollectionMethods'
+import { addReplyMethods } from './ReplyCollectionMethods'
 
 export const CommentsCollection = new Mongo.Collection('comments')
 
@@ -192,11 +193,13 @@ const calculateAverageRating = (ratings) => _.reduce(
   0
 )
 
+// TODO: put into lib and remove from here
 CommentsCollection._calculateAverageRating = calculateAverageRating
 
 const getCount = (scope, field) => scope[field] && scope[field].length ? scope[field].length : 0
 
 addCollectionMethods(CommentsCollection)
+addReplyMethods(CommentsCollection)
 
 CommentsCollection.helpers({
   likesCount: function () {
