+++ lib/lib/triggerEvent.js
@@ -1,0 +1,5 @@
+export const triggerEvent = (name, action, payload) => {
+  const func = Comments.config().onEvent
+
+  if (func) func(name, action, payload)
+}
