+++ lib/server/publish.js
@@ -64,9 +64,7 @@ const getCompositeCommentCursor = (rootCursor) => ({
 
 const getUserIdForPublication = (referenceId, userId, anonUserData) => {
   try {
-    userService.verifyAnonUserData(anonUserData, referenceId)
-
-    if (!userId && anonUserData._id) userId = anonUserData._id
+    userId = userService.verifyAnonUserData(anonUserData, referenceId)
   } catch(e) {}
 
   return userId
