+++ lib/services/reply.js
@@ -55,3 +55,9 @@ export const getMongoReplyFieldDescriptor = (replies, replyId) => {
     return acc
   }, { descriptor: '', found: false }).descriptor
 }
+
+export const allowReplies = referenceId => {
+  const config = Comments.config()
+
+  return config.allowReplies ? config.allowReplies(referenceId) : config.replies
+}
